Fix errors when compiling under GCC with pedantic errors on; all are comparisons between size_t and int32_t. (#158)

+++ cell/src/u_cell_sec.c
@@ -762,7 +762,7 @@ int32_t uCellSecPskGenerate(int32_t cellHandle,
                 }
                 uAtClientResponseStop(atHandle);
                 if ((uAtClientUnlock(atHandle) == 0) &&
-                    (sizeOutPsk == pskSizeBytes)) {
+                    (sizeOutPsk == (int32_t) pskSizeBytes)) {
                     errorCodeOrSize = sizeOutPskId;
                 }
             }
